wip: add del
